pyproject.toml
[project]
name = "teaparty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "google-auth>=2.40.0",
    "httpx>=0.28.0",
    "email-validator>=2.2.0",
    "anthropic>=0.52.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.32.0",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
]
